Revert "refactor frontend modules"
