+++ ExPlast/sitebuilder/frontend/index.html
@@ -130,8 +130,6 @@
 
   <input type="file" id="fileImport" accept=".json,.zip" style="display:none">
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jszip/3.10.1/jszip.min.js"></script>
-  <script type="module" src="builder.js"></script>
-  <script type="module" src="drag.js"></script>
-  <script type="module" src="props.js"></script>
+  <script src="builder.js" onload="builder.init()"></script>
 </body>
 </html>
